Refactored the controller tests to be in the api test package.
