+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/api/controllers/BackupImportController.java
@@ -1,6 +1,7 @@
-package com.sethhaskellcondie.thegamepensiveapi.api;
+package com.sethhaskellcondie.thegamepensiveapi.api.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sethhaskellcondie.thegamepensiveapi.api.FormattedResponseBody;
 import com.sethhaskellcondie.thegamepensiveapi.domain.backupimport.BackupDataDto;
 import com.sethhaskellcondie.thegamepensiveapi.domain.backupimport.BackupImportGateway;
 import com.sethhaskellcondie.thegamepensiveapi.domain.backupimport.ImportResultsDto;
