+++ src/main/java/com/sethhaskellcondie/thegamepensiveapi/api/controllers/HeartbeatController.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi.api;
+package com.sethhaskellcondie.thegamepensiveapi.api.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
