+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/api/controllers/HeartbeatControllerTests.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi.api;
+package com.sethhaskellcondie.thegamepensiveapi.api.controllers;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
