+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/backupimport/BackupImportGatewayTests.java
@@ -1,4 +1,4 @@
-package com.sethhaskellcondie.thegamepensiveapi.api;
+package com.sethhaskellcondie.thegamepensiveapi.domain.backupimport;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +10,7 @@ import org.springframework.test.web.servlet.MockMvc;
 @SpringBootTest
 @ActiveProfiles("test-container")
 @AutoConfigureMockMvc
-public class BackupImportControllerTests {
+public class BackupImportGatewayTests {
     @Autowired
     private MockMvc mockMvc;
